requirements.txt
python-dateutil
boto